pyproject.toml
[project]
name = "bambooML"
version = "0.1.0"
description = "Lightweight, modern ML/LLM framework with HEP-style data processing"
authors = [{name = "Xinzhu W"}]
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24,<2.0",
  "torch>=2.1,<3.0",
  "pyarrow>=12",
  "pandas>=2.0,<3.0",
  "PyYAML>=6,<7",
]

[project.scripts]
bambooml = "bambooml.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["bambooml*"]  

[project.optional-dependencies]
hep = [
  "uproot>=5,<6",
  "tables>=3.7,<4",
]
monitor = [
  "tensorboard>=2.13,<3",
]
llm = [
  "transformers>=4.38,<5",
  "peft>=0.10,<0.11",
  "accelerate>=0.23,<0.26",
]